--- src/mock-array.js
@@ -1,11 +1,0 @@
-import {getCard} from './data.js';
-
-export const mockArray = [];
-
-const addObjToArray = () => {
-  for (let i = 0; i < 22; i++) {
-    mockArray.push(getCard());
-  }
-};
-
-addObjToArray();
